Class discovery now relies on rescuing NameError.
This is far more sane than using regexes to parse the routes
file looking for class names. This method also allows us to
dynamically define methods that are called on, what I assume
are Rack apps, that return an object that responds to call.

Bumping the version to 0.1.0 b/c this is a major refactor.